@@ -33,7 +33,7 @@
       "artifactName": "${name}-${version}-setup.${ext}",
       "warningsAsErrors": false
     },
-    "productName": "Luminati Proxy Manager 1.93.58",
+    "productName": "Luminati Proxy Manager",
     "publish": {
       "provider": "generic",
       "url": "http://luminati.io/static/lpm"
@@ -97,17 +97,12 @@
     "compression": "1.7.2",
     "json-object-stream": "0.1.1",
     "socket-pair": "1.0.3",
-    "ws": "5.1.1"
+    "ws": "5.1.1",
+    "diacritic": "0.0.2"
   },
   "description": "A configurable local proxy for luminati.io",
   "devDependencies": {
-    "@uirouter/angularjs": "1.0.4",
-    "angular": "1.6.0",
-    "angular-google-analytics": "1.1.8",
-    "angular-sanitize": "1.6.0",
-    "angular-ui-bootstrap": "2.3.1",
     "animate.css": "3.5.2",
-    "axios": "0.16.2",
     "sinon": "4.2.0",
     "babel-cli": "6.24.1",
     "babel-loader": "7.1.1",
@@ -134,21 +129,19 @@
     "mocha": "3.2.0",
     "moment": "2.17.1",
     "nock": "9.0.11",
-    "react": "16.1.1",
+    "react": "16.3.2",
     "react-bootstrap": "0.31.0",
     "react-bootstrap-typeahead": "2.0.2",
     "react-dom": "16.1.1",
-    "react-autosuggest": "9.3.4",
-    "react-router": "4.2.0",
     "react-router-dom": "4.2.2",
     "style-loader": "0.18.2",
     "swagger-ui": "2.2.8",
-    "ui-select": "0.19.8",
     "url-loader": "0.5.9",
     "file-saver": "1.3.3",
     "webpack": "3.4.0",
     "recompose": "0.26.0",
-    "prop-types": "15.6.0"
+    "prop-types": "15.6.0",
+    "react-waypoint": "8.0.1"
   },
   "engines": {
     "node": ">=7.5",
@@ -280,6 +273,7 @@
       "bin/pub/img/zone_icon.png",
       "index.js",
       "lib/cities.js",
+      "lib/cluster_mode.js",
       "lib/cluster_route.js",
       "lib/db_history.json",
       "lib/find_iface.js",
@@ -298,21 +292,24 @@
       "src/pub/add_proxy.js",
       "src/pub/app.js",
       "src/pub/app.less",
+      "src/pub/chrome_widgets.js",
       "src/pub/common.js",
       "src/pub/config.js",
       "src/pub/edit_proxy.js",
+      "src/pub/har_preview.js",
+      "src/pub/har_viewer.js",
       "src/pub/howto.js",
       "src/pub/index.html",
       "src/pub/instructions.js",
       "src/pub/json_viewer.js",
       "src/pub/login.js",
       "src/pub/logs.js",
-      "src/pub/logs2.js",
       "src/pub/nav.js",
       "src/pub/notif_center.js",
       "src/pub/no_proxies.js",
       "src/pub/overview.js",
       "src/pub/proxies.js",
+      "src/pub/proxy_fields.js",
       "src/pub/proxy_tester.js",
       "src/pub/report_bug.js",
       "src/pub/schema.js",
@@ -321,7 +318,7 @@
       "test.js",
       "webpack.config.js"
     ],
-    "md5": "950c93d27702591702b3122f9bf0a205"
+    "md5": "e87b2260b9a940f707aa65b7d47890c2"
   },
   "main": "index.js",
   "name": "@luminati-io/luminati-proxy",
@@ -362,5 +359,5 @@
     "test-win": "node node_modules/mocha/bin/mocha",
     "build": "node_modules/webpack/bin/webpack.js"
   },
-  "version": "1.93.58"
+  "version": "1.95.360"
 }
