@@ -5,75 +5,37 @@ import Proxies from './proxies.js';
 import ajax from 'hutil/util/ajax';
 import setdb from 'hutil/util/setdb';
 import Stats from './stats.js';
-import Logs from './logs';
+import Har_viewer from './har_viewer';
 import Pure_component from '../../www/util/pub/pure_component.js';
 import {If} from '/www/util/pub/react.js';
-import {is_electron, Loader} from './common.js';
-import zurl from 'hutil/util/url';
 import $ from 'jquery';
 
 class Overview extends Pure_component {
-    state = {loading: false};
-    componentWillMount(){
-        this.setdb_on('head.section', section=>{
-            if (!section)
-                return;
-            const update = {section};
-            if (section.name=='overview_multiplied')
-            {
-                update.master_port = window.location.pathname.split('/')
-                .slice(-1)[0];
-            }
-            this.setState(update);
-        });
+    componentDidMount(){
         this.setdb_on('head.proxies_running', proxies=>
             this.setState({proxies}));
         this.setdb_on('head.consts', consts=>this.setState({consts}));
-        window.setTimeout(this.fetch_globals.bind(this));
-    }
-    fetch_globals(){
-        const _this = this;
-        this.etask(function*(){
-            // XXX krzysztof: optimize /api/consts endpoint
-            //_this.setState({loading: true});
-            if (!_this.state.consts)
-            {
-                const consts = yield ajax.json({url: '/api/consts'});
-                setdb.set('head.consts', consts);
-            }
-            _this.setState({loading: false});
-        });
     }
     render(){
-        const title = this.state.master_port ?
-            `Overview of multiplied port - ${this.state.master_port}` :
-            'Overview';
-        let ports;
-        if (this.state.master_port&&this.state.proxies)
-        {
-            const mp = this.state.proxies.find(p=>
-                p.port==[this.state.master_port]);
-            ports = [];
-            for (let i=mp.port; i<mp.port+mp.multiply; i++)
-                ports.push(i);
-        }
+        const master_port = this.props.match.params.master_port;
+        const title = master_port ?
+            `Overview of multiplied proxy port - ${master_port}` : 'Overview';
         return (
-            <div className="overview lpm">
-              <Loader show={this.state.loading}/>
+            <div className="overview_page lpm">
               <Upgrade/>
               <div className="proxies nav_header">
                 <h3>{title}</h3>
               </div>
               <div className="panels">
                 <div className="proxies proxies_wrapper">
-                  <Proxies master_port={this.state.master_port}/>
+                  <Proxies master_port={master_port}/>
                 </div>
                 <div className="stats_wrapper">
-                  <Stats master_port={this.state.master_port}/>
+                  <Stats master_port={master_port}/>
                 </div>
               </div>
               <div className="logs_wrapper">
-                <Logs ports={ports}/>
+                <Har_viewer master_port={master_port}/>
               </div>
             </div>
         );
@@ -96,6 +58,7 @@ class Upgrade extends Pure_component {
     render(){
         const {upgrading, upgrade_error, ver_last, ver_node} = this.state;
         const is_upgradable = ver_last&&ver_last.newer;
+        const is_electron = window.process && window.process.versions.electron;
         const electron = ver_node&&ver_node.is_electron||is_electron;
         if (!is_upgradable||!ver_node)
             return null;
