@@ -16,11 +16,9 @@ E.is_env_compat = ()=>{
     }
     const path = require('path');
     const pkg = require('../package.json');
-    const excluded = ['angular', 'angular-sanitize', 'bootstrap',
-        'bootstrap-datepicker', 'codemirror', 'notosans-fontface',
-        'require-css', 'flag-icon-css', 'angular-ui-bootstrap',
-        'swagger-ui', 'font-awesome', 'angular-google-analytics', 'ui-select',
-        '@uirouter/angularjs', 'animate.css']
+    const excluded = ['bootstrap', 'bootstrap-datepicker', 'codemirror',
+        'notosans-fontface', 'require-css', 'flag-icon-css', 'swagger-ui',
+        'font-awesome', 'animate.css']
         .concat(Object.keys(pkg.optionalDependencies));
     for (let dep in pkg.dependencies)
     {
