@@ -1,6 +1,5 @@
 // LICENSE_CODE ZON
 'use strict'; /*jslint node:true*/
-require('./config.js');
 const net = require('net');
 const cluster = require('cluster');
 const child_process = require('child_process');
@@ -12,15 +11,11 @@ const socket_pair = require('socket-pair');
 const {etask, zerr, conv, array} = hutil;
 const zutil = hutil.util;
 const hash = require('./hash.js');
-// const znet = require('./net.js');
-// const proc = require('./proc.js');
-// const toobusy = require('./toobusy.js');
+const os = require('os');
 
-const znet = {};
 const proc = {
-    cluster_workers_n: ()=>{ return 2; }
+    cluster_workers_n: ()=>{ return os.cpus().length; },
 };
-const toobusy = {};
 
 const E = exports, assign = Object.assign;
 zutil.inherit_init(E, EventEmitter);
@@ -103,6 +98,8 @@ let worker_init = channel_needed=>etask(function*worker_init(){
         if (!socket._handle||!other._handle)
             throw new Error(`cluster_route: bad ipc socket`);
         tx_w2w_channel[i] = E._create_ipc_channel(socket._handle.fd);
+        zerr('cluster_route: w2w_channel, to:%s, from:%s', i,
+            cluster.worker.id);
         process.send({cmd: 'cluster_route:w2w_channel', to_worker_id: i,
             from_worker_id: cluster.worker.id}, other._handle);
     }
@@ -113,6 +110,7 @@ let worker_init = channel_needed=>etask(function*worker_init(){
             zerr('cluster_route no handle: %O', msg);
         let rx_w2w_channel = E._create_ipc_channel(recv_handle.fd);
         rx_w2w_channel.on('message', (w2w_msg, w2w_recv_handle)=>{
+            console.log('cluster_route: channel message', w2w_msg, w2w_recv_handle.fd);
             if (w2w_msg && w2w_msg.cmd=='cluster_route:connection')
             {
                 if (!w2w_recv_handle)
@@ -179,9 +177,11 @@ let pipe_wrap_socket = src_socket=>etask(function*pipe_wrap_socket(){
     return ret_socket;
 });
 E.send = (msg, send_handle, to_worker_id)=>etask(function*send(){
+    console.log('cluster_route: send', msg, send_handle._handle.fd, to_worker_id, to_worker_id==cluster.worker.id);
     if (to_worker_id==cluster.worker.id)
         return void E.emit('message', msg, send_handle, cluster.worker.id);
     let w2w_channel;
+    console.log('send, has_channel', !!tx_w2w_channel[to_worker_id]);
     if (!(w2w_channel = tx_w2w_channel[to_worker_id]))
         throw new Error(`No channel to worker '${to_worker_id}'`);
     if (send_handle)
@@ -189,9 +189,11 @@ E.send = (msg, send_handle, to_worker_id)=>etask(function*send(){
         msg.send_handle_src = {ip: send_handle.remoteAddress,
             port: send_handle.remotePort, ssl: send_handle.encrypted};
     }
+    console.log('send send_handle_src', msg.send_handle_src);
     if (send_handle && send_handle.encrypted)
     {
         send_handle._handle.readStart();
+        console.log('send wrap socket');
         send_handle = yield pipe_wrap_socket(send_handle);
     }
     if (msg.cmd=='cluster_route:connection')
@@ -201,6 +203,7 @@ E.send = (msg, send_handle, to_worker_id)=>etask(function*send(){
         else
             msg.no_handle = true;
     }
+    console.log('send w2w channel send', msg, send_handle._handle.fd);
     w2w_channel.send(msg, send_handle);
 });
 let set_hook = (ee, event_name, hook_fn)=>{
@@ -220,16 +223,18 @@ let select_worker_by_ip = socket=>
 let select_worker_by_ip_smart = (socket, req)=>ip_hash(
     req.headers['x-forwarded-for']||socket.remoteAddress)
     %proc.cluster_workers_n()+1;
-let select_worker_by_port = (socket, req)=>hash.hash_int(socket.remotePort)%proc.cluster_workers_n()+1;
+
+let select_worker_by_server_id = (socket, req)=>{
+    let server_id = socket.route.opt.server_id||"none";
+    return hash.hash_string(server_id)%proc.cluster_workers_n()+1;
+};
 let check_fallback = to_worker_id=>{
+    console.log('check_fallback start', to_worker_id);
     if (!to_worker_id)
         to_worker_id = cluster.worker.id;
-    if (to_worker_id==cluster.worker.id
-        && toobusy.actions.cluster_route_fallback
-        && proc.is_worker_busy(to_worker_id))
-    {
+    if (to_worker_id==cluster.worker.id)
         to_worker_id = cluster.worker.id;
-    }
+    console.log('check_falback end', to_worker_id, cluster.worker.id);
     return to_worker_id;
 };
 let conn_hook = (pass_conn, socket)=>void etask(function*_conn_hook(){
@@ -338,6 +343,8 @@ E.setup = (server, opt)=>{
         assign(opt, {select: select_worker_by_ip_smart, need_headers: true});
     else if (opt.select=='ip')
         assign(opt, {select: select_worker_by_ip});
+    else if (opt.select=='server_id')
+        assign(opt, {select: select_worker_by_server_id});
     let conn_ev = server instanceof tls.Server && opt.need_headers
         ? 'secureConnection' : 'connection';
     if (opt.need_headers)
@@ -345,20 +352,18 @@ E.setup = (server, opt)=>{
     else
         set_hook(server, conn_ev, conn_hook);
     E.on('cluster_route:connection', (msg, socket)=>{
+        console.log('on connection1', msg, opt.server_id, socket._handle.fd);
         if (msg.server_id!=opt.server_id)
             return;
-        if (opt.check_busy!=false && toobusy.actions.cluster_route_disconnect
-            && toobusy.is_busy())
-        {
-            return socket.destroy();
-        }
         socket.route = {src: msg.send_handle_src, routed: true};
         let src = socket.route.src;
         if (src && src.ip)
             socket._peername = {address: src.ip, port: src.port, family: 4};
         server.emit(conn_ev, socket);
+        console.log('on connection2', msg.consumed_data);
         if (msg.consumed_data)
             socket.unshift(new Buffer(msg.consumed_data, 'base64'));
+        socket.resume();
     });
     set_hook(server, conn_ev, (pass_conn, socket)=>{
         socket.route = assign({opt}, socket.route);
@@ -369,6 +374,6 @@ E.setup = (server, opt)=>{
 E.init = opt=>{
     opt = opt||{};
     if (cluster.isMaster)
-        return void master_init(opt.external);
-    worker_init(opt.channel_needed, opt.external);
+        return void master_init();
+    worker_init(opt.channel_needed);
 };
