@@ -1,14 +1,7 @@
 // LICENSE_CODE ZON
 'use strict'; /*jslint node:true, browser:true*/
-(function(){
-var define;
-var is_node_ff = typeof module=='object' && module.exports;
-if (!is_node_ff)
-    define = self.define;
-else
-    define = require('./require_node.js').define(module, '../');
-define([], function(){
-var E = {};
+
+var E = module.exports = {};
 
 function imul64h(a, b){
     var ah = (a>>>16)&0xffff;
@@ -45,5 +38,3 @@ E.sum_parse = function(str){
     return {hash: parts[1], type: parts[2]=='*' ? 'binary' : 'text',
         file: parts[3]};
 };
-
-return E; }); }());
