@@ -1132,7 +1132,6 @@ E.prototype._aggregate_history = function(response){
     data.in_bw = response.in_bw;
     data.out_bw = response.out_bw;
     this.opt.history_aggregator(data);
-
 };
 
 E.prototype.get_other_port = function(port){
@@ -1222,12 +1221,15 @@ E.prototype.error_handler = function error_handler(source, err){
     if (this.is_overload(err))
         return;
     this.log.error(source+' error '+zerr.e2s(err));
-    if (err.code=='EADDRINUSE')
+    let {code} = err;
+    if (code=='EADDRINUSE')
     {
-        err = new Error('There\'s already an application which runs on '
-        +err.address+':'+err.port);
+        err = new Error(`There's already an application which runs on `+
+            `${err.address}:${err.port}`);
         err.raw = true;
+        err.code = code;
     }
+    err.lum_source = source;
     if (this.listenerCount('error'))
         this.emit('error', err);
     else
@@ -1786,14 +1788,6 @@ E.prototype._handle_proxy_resp = function(req, res, proxy, task, head,
         try {
             ctx.timeline.track('response');
             let code = `${_res.statusCode}`.replace(/(?!^)./g, 'x');
-            let proxy_err = this._check_proxy_response(ctx.host, _res,
-                'response');
-            if (proxy_err)
-            {
-                req.ctx.log.warn('proxy api err %s: %s', proxy_err.code,
-                    proxy_err.message);
-                return task.throw(proxy_err);
-            }
             ctx.log.info(`${req.method} ${ctx.url} - ${_res.statusCode}`);
             if (ctx.session)
             {
@@ -2018,7 +2012,6 @@ E.prototype._handle_proxy_error= function(req, res, proxy, task, head){
             this._abort_proxy_req(req, proxy);
             return task.throw(new Error('overload'));
         }
-        // XXX maximk: check if X-Luminati-Error is caught here
         let proxy_err = this._check_proxy_response(ctx.host,
             res||{statusCode: 502}, {from: 'error', error: err});
         let can_retry = ctx.rules && ctx.rules.can_retry(req, ctx.response, {
