@@ -15,7 +15,6 @@ const crypto = require('crypto');
 const express = require('express');
 const compression = require('compression');
 const body_parser = require('body-parser');
-const json2csv = require('json2csv');
 const log = require('./log.js');
 const http_shutdown = require('http-shutdown');
 const Luminati = require('./luminati.js');
@@ -25,7 +24,6 @@ const find_iface = require('./find_iface.js');
 const pkg = require('../package.json');
 const swagger = require('./swagger.json');
 const request = require('request').defaults({gzip: true});
-const sequelize = require('sequelize');
 const http = require('http');
 const https = require('https');
 const hutil = require('hutil');
@@ -48,6 +46,7 @@ const perr = require('./perr.js');
 const {Netmask} = require('netmask');
 const util_lib = require('./util.js');
 const web_socket = require('ws');
+const cluster_mode = require('./cluster_mode.js');
 
 let phantomjs, webshot, posix, cookie_jar, proxy_creds_check_cache = {};
 try {
@@ -241,7 +240,8 @@ function Manager(args, run_config){
     this.argv.log = this.argv.log.toLowerCase();
     this._log = log(this.argv.www||'MNGR', this.argv.log,
         this.argv.disable_color);
-    this._log.notice('Manager started %s', pkg.version);
+    if (cluster_mode.is_master())
+        this._log.notice('Manager started %s', pkg.version);
     if (this.argv.session=='true')
         this.argv.session = true;
     this._log.debug('Effective args %O', this.argv);
@@ -272,43 +272,36 @@ function Manager(args, run_config){
             {
                 return;
             }
-            // XXX krzysztof: stat is to be removed
-            let stat = _.pick(data, qw`response_headers request_headers
-                status_message response_time url proxy_peer success timeline
-                method content_size port status_code`);
-            stat.response_body = data.response_body.slice(0, 2048);
+            if (data.response_body.type)
+                data.response_body = Buffer.from(data.response_body.data);
+            data.response_body = data.response_body.slice(0, 2048);
             let url_parts;
             if (url_parts = data.url.match(/^([^\/]+?):(\d+)$/))
             {
-                stat.protocol = url_parts[2]==443 ? 'https' : 'http';
-                stat.hostname = url_parts[1];
+                data.protocol = url_parts[2]==443 ? 'https' : 'http';
+                data.hostname = url_parts[1];
             }
             else
             {
-                assign(stat, _.pick(url.parse(data.url),
+                assign(data, _.pick(url.parse(data.url),
                     qw`protocol hostname`));
-                stat.protocol = (stat.protocol||'https:').slice(0, -1);
+                data.protocol = (data.protocol||'https:').slice(0, -1);
             }
-            if (stat.hostname)
-            {
-                stat.in_bw = data.in_bw;
-                stat.out_bw = data.out_bw;
-            }
-            data.protocol = stat.protocol;
-            data.hostname = stat.hostname;
             data.success = +(data.status_code &&
                 /([23]..|404)/.test(data.status_code));
+            data.content_type = get_content_type(data);
             setImmediate(()=>this.handle_usage(data));
             let proxy;
             const full_logs = this.argv.history ||
             (proxy = this.proxies_running[data.port]) && proxy.opt.history;
             setImmediate(()=>this.history_aggregator(data, {full_logs}));
         } catch(e){
-            // XXX maximk: buffer failed aggregations
             this._log.error('stats aggregator %s', zerr.e2s(e));
         }
     };
     this.handle_usage = etask._fn(function*mgr_handle_usage(_this, data){
+        if (!data.url||data.url.match(/lumtest\.com\/myip\.json/))
+            return;
         try {
             if (!data.out_bw && !data.in_bw)
                 return;
@@ -334,22 +327,23 @@ function Manager(args, run_config){
                 ${curr.success+data.success}, '${data.status_code}',
                 '${data.hostname}', '${data.protocol}')`);
         } catch(e){
-            // XXX maximk: buffer failed aggregations
             this._log.error('handle usage %s', zerr.e2s(e));
         }
     });
+    this.worker_aggregator = data=>{
+        process.send({cmd: 'w2m_record_stats', data});
+    };
     this.history_aggregator = (data, opt={})=>{
         try {
+            if (data.response_body.type)
+                data.response_body = Buffer.from(data.response_body.data);
             data.response_body = data.response_body.toString();
-            const har_req = this.har([data]).log.entries[0];
-            this.wss.broadcast(JSON.stringify(har_req));
             const _this = this;
             etask(function*history_aggregator(){
                 try {
                     if (!opt.full_logs)
                     {
                         let ports = [..._this.stats_proxies].join(',');
-                        let delete_limit = 10;
                         let limit = _this._defaults.request_stats_limit;
                         const curr_count = yield _this.sql(`SELECT COUNT(*) AS
                             count FROM request WHERE port IN (${ports});`);
@@ -369,13 +363,16 @@ function Manager(args, run_config){
                     let row = {};
                     for (let f in data)
                         row['$'+f] = data[f];
-                    _this.db.main.stmt.history.run(row);
-                } catch (e){
+                    _this.db.main.stmt.history.run(row, function(err){
+                        data.uuid = this.lastID;
+                        const har_req = _this.har([data]).log.entries[0];
+                        _this.wss.broadcast(JSON.stringify(har_req));
+                    });
+                } catch(e){
                     _this._log.error('history aggregator %s', zerr.e2s(e));
                 }
             });
         } catch(e){
-            // XXX maximk: buffer failed aggregations
             this._log.error('history aggregator %s', zerr.e2s(e));
         }
     };
@@ -397,6 +394,29 @@ function Manager(args, run_config){
         }
         ua.event('manager', 'use_as_module').send();
     }
+    cluster_mode.add_msg_hdlr(this.handle_message.bind(this));
+}
+
+function get_content_type(data){
+    let content_type;
+    let res = 'other';
+    try {
+        const headers = JSON.parse(data.response_headers);
+        content_type = headers['content-type']||'';
+    } catch(e){ content_type = ''; }
+    if (content_type.match(/json/))
+        res = 'xhr';
+    else if (content_type.match(/javascript/))
+        res = 'js';
+    else if (content_type.match(/css/))
+        res = 'css';
+    else if (content_type.match(/image/))
+        res = 'img';
+    else if (content_type.match(/audio|video/))
+        res = 'media';
+    else if (content_type.match(/font/))
+        res = 'font';
+    return res;
 }
 
 function _config_from_file(fname, proxy_type){
@@ -483,6 +503,18 @@ function _get_proxy_configs(_this, args){
 
 util.inherits(Manager, events.EventEmitter);
 
+E.prototype.handle_message = function(wkr, msg){
+    let {cmd, data} = msg;
+    if (cmd == 'w2m_record_stats')
+        return this.record_stats(data);
+    if (cmd == 'm2w_refresh_sessions')
+        return this.refresh_server_sessions(data);
+    if (cmd == 'm2w_proxy_delete')
+        return this.proxy_delete(data);
+    if (cmd == 'm2w_proxy_create')
+        return this.proxy_create(data);
+};
+
 E.prototype._api = function(f, roles){
     const _this = this;
     return (req, res, next)=>etask(function*mgr__api(){
@@ -522,9 +554,13 @@ function*mgr_stop_servers(_this, force, www){
     }));
     if (www && _this.www_server)
         stop_server(_this.www_server);
-    _.values(_this.socks_servers).forEach(stop_server);
-    _.values(_this.proxies_running).forEach(stop_server);
-    _this.wss.close();
+    if (cluster_mode.is_worker())
+    {
+        _.values(_this.socks_servers).forEach(stop_server);
+        _.values(_this.proxies_running).forEach(stop_server);
+    }
+    if (_this.wss)
+        _this.wss.close();
     yield etask.all(servers);
     _this._log.debug('Servers stopped');
 });
@@ -547,18 +583,22 @@ E.prototype.stop_dbs = etask._fn(function*mgr_stop_dbs(_this){
 
 E.prototype.stop = etask._fn(
 function*mgr_stop(_this, reason, force, restart){
-    yield _this.perr(restart ? 'restart' : 'exit', {reason});
     _this.is_running = false;
+    if (cluster_mode.is_master())
+    {
+        yield _this.perr(restart ? 'restart' : 'exit', {reason});
+        yield _this.stop_dbs();
+        if (reason!='config change')
+            yield _this.save_config();
+        if (reason instanceof Error)
+            reason = zerr.e2s(reason);
+        _this._log.notice('Manager stopped %O', {reason, force, restart});
+    }
     yield _this.stop_servers(force, true);
-    yield _this.stop_dbs();
-    if (reason!='config change')
-        yield _this.save_config();
     if (_this.ulimit_mgr)
         _this.ulimit_mgr.stop();
     _this.rmt_cfg.stop();
-    if (reason instanceof Error)
-        reason = zerr.e2s(reason);
-    _this._log.notice('Manager stopped %O', {reason, force, restart});
+
     if (!restart)
         _this.emit('stop', reason);
 });
@@ -602,6 +642,9 @@ E.prototype.har = function(entries){
                     proxy_peer: entry.proxy_peer,
                     protocol: entry.protocol,
                     port: entry.port,
+                    timestamp: entry.timestamp,
+                    content_type: entry.content_type,
+                    success: entry.success,
                 },
                 startedDateTime: new Date(tl.create).toISOString(),
                 time: elapsed+blocked,
@@ -647,6 +690,8 @@ E.prototype.har = function(entries){
 E.prototype.save_config = function(){
     let filename = this.argv.config;
     if (!filename || !_.isString(filename))
+        return;
+    if (!cluster_mode.is_master())
         return;
     let proxies = this.proxies
         .filter(p=>p.proxy_type=='persist')
@@ -779,6 +824,8 @@ function*mgr_prepare_database(_this){
         sqlite3.verbose();
     _this.db = {};
     _this.db.main = yield open_db(_this.argv.database);
+    if (cluster_mode.is_worker(true))
+        return;
     let now = Date.now();
     let rows = yield _this.sql(
         'SELECT name FROM sqlite_master WHERE type="table"');
@@ -847,7 +894,29 @@ function*mgr_prepare_database(_this){
     let archive_timestamps = yield _this.archive_timestamps();
     for (let i in archive_timestamps)
         yield _this.archive_copy(archive_timestamps[i], !i);
+
 });
+
+E.prototype.prepare_db_stmt = function(){
+    this.db.main.stmt = {
+        proxy_remove: this.db.main.prepare(`DELETE FROM proxy_hosts
+            WHERE host=$host`),
+        history: this.db.main.prepare(`INSERT INTO request (port, url,
+            method, request_headers, request_body, response_headers,
+            status_code, status_message, timestamp, elapsed,
+            response_time, node_latency, country, timeline,
+            super_proxy, proxy_peer, username, content_size, context,
+            success, out_bw, in_bw,response_body, protocol, hostname,
+            content_type)
+            VALUES ($port, $url, $method, $request_headers,
+            $request_body, $response_headers, $status_code,
+            $status_message, $timestamp, $elapsed, $response_time,
+            $node_latency, $country, $timeline, $super_proxy,
+            $proxy_peer, $username, $content_size, $context,
+            $success, $out_bw, $in_bw, $response_body, $protocol,
+            $hostname, $content_type)`),
+    };
+};
 
 E.prototype.get_zones = etask._fn(function*mgr_get_zones(_this){
     let config = yield _this.get_lum_local_conf();
@@ -1007,6 +1076,7 @@ E.prototype.error_handler = function error_handler(source, err){
     if (!err.raw)
         this._log.error(source+' error '+zerr.e2s(err));
     console.log('error_handler', err, zerr.e2s(err));
+    err.source = source;
     this.emit('error', err);
 };
 
@@ -1074,14 +1144,22 @@ function*mgr_create_single_proxy(_this, conf){
     conf = _this._complete_proxy_config(conf);
     let server = new Luminati(conf, _this);
     server.on('error', err=>_this.error_handler('Proxy '+conf.port, err));
-    _this._log.notice('Starting proxies %s', conf.socks ?
-        `${conf.port}:${conf.socks}` : conf.port);
-    yield server.listen();
+    if (cluster_mode.is_master())
+    {
+        _this._log.notice('Starting proxies %s', conf.socks ?
+            `${conf.port}:${conf.socks}` : conf.port);
+    }
+    if (cluster_mode.is_worker())
+    {
+        yield server.listen();
+        if (server.http_server)
+            cluster_mode.server_wrap(server.http_server);
+    }
     _this._log.debug('local proxy %O', _.omit(server.opt, qw`proxy_cache
         history_aggregator`));
-    _this.proxies_running[server.port] = server;
+    _this.proxies_running[server.opt.listen_port] = server;
     if (!server.opt.history)
-        _this.stats_proxies.add(server.port);
+        _this.stats_proxies.add(server.opt.listen_port);
     if (conf._update)
     {
         setImmediate(()=>{
@@ -1090,6 +1168,15 @@ function*mgr_create_single_proxy(_this, conf){
         });
     }
     return server;
+});
+
+E.prototype.record_stats = etask._fn(function*mgr_record_stats(_this, stats){
+    if (!stats||!stats.port)
+        return;
+    let proxy = _this.proxies_running[stats.port];
+    if (!proxy||!proxy.opt.history_aggregator)
+        return;
+    yield proxy.opt.history_aggregator(stats);
 });
 
 const nop = ()=>{};
@@ -1102,8 +1189,9 @@ function*mgr_create_proxy(_this, proxy){
     let conf =
         assign({
             proxy_cache: _this.proxy_cache,
-            history_aggregator: _this.argv.request_stats ?
-                _this.stats_aggregator : _this.history_aggregator,
+            history_aggregator: cluster_mode.is_worker(true)
+                ? _this.worker_aggregator : _this.argv.request_stats
+                ? _this.stats_aggregator : _this.history_aggregator,
             proxy_internal_bypass: PROXY_INTERNAL_BYPASS,
         }, proxy);
     if (_this.argv.request_stats)
@@ -1120,10 +1208,17 @@ function*mgr_create_proxy(_this, proxy){
     });
     server.stop = ()=>etask(function*mgr_server_stop(){
         const args = [].slice.call(arguments);
-        _this._log.notice('Stopping proxies %s', servers.map(s=>s.port));
+        if (cluster_mode.is_master())
+        {
+            _this._log.notice('Stopping proxies %s', servers.map(s=>s.port||
+            s.opt.listen_port||s.opt.port));
+        }
         return yield etask.all(servers.map(s=>{
-            delete _this.proxies_running[s.port];
-            _this.stats_proxies.delete(s.port);
+            let port = s.port||s.opt.listen_port||s.opt.port;
+            delete _this.proxies_running[port];
+            _this.stats_proxies.delete(port);
+            if (cluster_mode.is_master(true))
+                return;
             return Luminati.prototype.stop.apply(s, args);
         }));
     });
@@ -1210,6 +1305,8 @@ E.prototype.proxy_create = etask._fn(function*mgr_proxy_create(_this, data){
         _this.proxies.push(proxy);
         _this.save_config();
     }
+    if (cluster_mode.is_master(true))
+        cluster_mode.broadcast({cmd: 'm2w_proxy_create', data});
     return server;
 });
 
@@ -1232,6 +1329,8 @@ E.prototype.proxy_delete = etask._fn(function*mgr_proxy_delete(_this, port){
         _this.proxies.splice(idx, 1);
         _this.save_config();
     }
+    if (cluster_mode.is_master(true))
+        cluster_mode.broadcast({cmd: 'm2w_proxy_delete', data: port});
 });
 
 const get_free_port = proxies=>{
@@ -1351,8 +1450,19 @@ E.prototype.refresh_sessions = function(req, res){
     let server = this.proxies_running[port];
     if (!server)
         return res.status(400, 'Invalid proxy port').end();
-    server.session_mgr.refresh_sessions();
+    if (cluster_mode.is_master(true))
+        cluster_mode.broadcast({cmd: 'm2w_refresh_sessions', data: port});
+    else
+        this.refresh_server_sessions(port);
     res.status(204).end();
+};
+
+E.prototype.refresh_server_sessions = function(port){
+    let server = this.proxies_running[port];
+    if (!server)
+        return false;
+    server.session_mgr.refresh_sessions();
+    return true;
 };
 
 E.prototype.proxy_status_get = etask._fn(
@@ -1901,7 +2011,7 @@ function*logs_suggestions(_this, req, res){
     const suggestions = {
         ports: ports.map(v=>v.val),
         domains: domains.map(v=>v.val),
-        codes: codes.map(v=>v.val),
+        status_codes: codes.map(v=>v.val),
         protocols: protocols.map(v=>v.val),
     };
     res.json(suggestions);
@@ -1931,17 +2041,23 @@ E.prototype.filtered_get = etask._fn(function*mgr_filtered_get(_this, req){
     const for_db = s=>(s||'').replace("'", "''");
     const skip = +req.query.skip||0;
     const limit = +req.query.limit||0;
-    req.query.status_code = req.query.code;
-    req.query.hostname = req.query.domain;
-    const wheres = ['port', 'hostname', 'protocol', 'status_code'].map(param=>{
-        if (!req.query[param])
-            return null;
-        return req.query[param].split(',').map(v=>`${param} = '${for_db(v)}'`)
-            .join(' OR ');
-    }).filter(Boolean).join(') AND (');
-    const where = wheres ? 'WHERE ('+wheres+')' : '';
-    const order_by = req.query.sort ? `ORDER BY ${req.query.sort}
-        ${req.query.sort_desc ? 'DESC' : ''}` : '';
+    const wheres = [];
+    if (req.query.port_from && req.query.port_to)
+    {
+        wheres.push(`port >= ${req.query.port_from} AND port <=
+            ${req.query.port_to}`);
+    }
+    if (req.query.search)
+        wheres.push(`url LIKE '%${for_db(req.query.search)}%'`);
+    ['port', 'content_type', 'status_code', 'protocol'].forEach(param=>{
+        let val;
+        if (val = req.query[param])
+            wheres.push(`${param} = '${for_db(val)}'`);
+    });
+    const where = wheres.length ? 'WHERE ('+wheres.join(') AND (')+')' : '';
+    const order_by = 'ORDER BY'+(req.query.sort ?
+        ` ${req.query.sort} ${req.query.sort_desc ? 'DESC' : ''},` : '')
+        +` uuid ${req.query.sort_desc ? 'DESC' : ''}`;
     const items = yield _this.sql(
         `SELECT in_bw+out_bw as bw, * FROM request ${where} ${order_by}`
         + (limit ? ` LIMIT ${limit} OFFSET ${skip}` : ''));
@@ -2362,13 +2478,13 @@ function*mgr_recent_stats(_this, req, res){
     let https;
     if (!(https = stats.protocol.find(p=>p.key=='https')))
     {
-        stats.protocol.push({key: 'https', out_bw: 0, in_bw: 0, reqs: 0,
-            enable});
+        stats.protocol.push({key: 'https', out_bw: 0, in_bw: 0, reqs: 0});
+        stats.ssl_enable = enable;
     }
     else if (https.reqs>0)
     {
-        https.warning = enable;
-        https.enable = enable;
+        stats.ssl_warning = enable;
+        stats.ssl_enable = enable;
     }
     const general = yield _this.sql(`SELECT SUM(success) as success,
         SUM(reqs) as total FROM usage WHERE port_id ${ports_cond}`);
@@ -2483,7 +2599,7 @@ function*mgr_create_web_interface(_this){
         res.locals.path = req.path;
         next();
     });
-    app.get(/^\/((|login|howto|proxy_tester|logs|logs2|config)|(overview|status_codes|domains|protocols|proxy)(\/[^\/]+)?)$/,
+    app.get(/^\/((|login|howto|proxy_tester|logs|dock_logs|config)|(overview|status_codes|domains|protocols|proxy)(\/[^\/]+)?)$/,
         _this._api(
         (req, res, next)=>{
             setTimeout(()=>etask(function*mgr_cities_ensure_data(){
@@ -2499,8 +2615,6 @@ function*mgr_create_web_interface(_this){
         }
     ));
     app.use(express.static(path.resolve(__dirname, '../bin/pub')));
-    app.use('/uib', express.static(path.resolve(__dirname,
-        '../node_modules/angular-ui-bootstrap')));
     app.use('/req', express.static(path.resolve(__dirname,
         '../node_modules')));
     app.use((req, res, next)=>{
@@ -2561,7 +2675,8 @@ function*mgr_create_socks_servers(_this){
     function*mgr_create_socks_server(){
         let port, server = new Socks(opt)
         .on('error', err=>_this.error_handler('Socks '+port, err));
-        yield server.start();
+        if (cluster_mode.is_worker())
+            yield server.start();
         port = server.port;
         _this.socks_servers[port] = server;
         _this._log.debug('SOCKS5 is available at port %s', port);
@@ -2756,6 +2871,11 @@ function*mgr_check_user(_this){
 E.prototype.login_user = etask._fn(
 function*mgr_login_user(_this, token, username, password, customer){
     let config, login_failed;
+    let www_api;
+    if (_this.current_country && _this.current_country.toLowerCase()!='cn')
+        www_api = 'https://luminati.io';
+    else
+        www_api = 'https://luminati-china.io';
     try { config = yield _this.get_lum_local_conf(customer, token, true); }
     catch(e){
         if (!e.status)
@@ -2785,7 +2905,7 @@ function*mgr_login_user(_this, token, username, password, customer){
                     message: `The email address is not registered. `
                         +`If you signed up with Google signup button, you`
                         +` should login with Google login button.`
-                        +` <a href="${_this.argv.api}/?need_signup=1"`
+                        +` <a href="${www_api}/?need_signup=1"`
                         +` target=_blank>`
                         +`Click here to sign up.</a>`,
                     email: true,
@@ -2793,7 +2913,7 @@ function*mgr_login_user(_this, token, username, password, customer){
             }
             return {error: {
                 message: `The password is incorrect. `
-                    +`<a href="${_this.argv.api}/forgot_password?email=`
+                    +`<a href="${www_api}/forgot_password?email=`
                     +`${encodeURIComponent(username)}" target=_blank>`
                     +`Forgot your password?</a>`,
                 password: true,
@@ -2805,8 +2925,7 @@ function*mgr_login_user(_this, token, username, password, customer){
     try {
         if (login_failed)
             config = yield _this.get_lum_local_conf(customer, token, true);
-    }
-    catch(e){
+    } catch(e){
         if (!e.status)
             throw e;
         if (e.status==403 && (e.message=='Your account is not active' ||
@@ -2827,7 +2946,7 @@ function*mgr_login_user(_this, token, username, password, customer){
             if (msg=='Your account is not active')
             {
                 msg = `Your account is disabled.`
-                +`<a href='${_this.argv.api}/cp/billing'>`
+                +`<a href='${www_api}/cp/billing'>`
                 +`Click here to change your account status</a>`;
             }
             return {error: {
@@ -2850,41 +2969,41 @@ function*mgr_login_user(_this, token, username, password, customer){
         {customers: config.customers.sort()} : {defaults: config._defaults};
 });
 
+E.prototype.set_location = etask._fn(function*mgr_set_location(_this){
+    let res = yield etask.nfn_apply(request, [{
+        url: 'http://lumtest.com/myip.json',
+        json: true,
+        timeout: 60*date.ms.SEC,
+    }]);
+    _this.current_country = res.body.country;
+});
+
 E.prototype.start = etask._fn(function*mgr_start(_this){
     this.on('uncaught', e=>_this.log.error('start %s', zerr.e2s(e)));
     try {
         perr.run({no_usage_stats: _this.argv.no_usage_stats});
+        yield _this.set_location();
         yield _this.rmt_cfg.start();
-        let errors = _this.config_check({_defaults: _this.argv,
-            proxies: _this.proxies});
-        if (errors.length)
+        if (cluster_mode.is_master())
         {
-            _this.emit('error', {message: 'Config errors', errors: errors});
-            throw new Error('Config errors: ' + JSON.stringify(errors));
+            let errors = _this.config_check({_defaults: _this.argv,
+                proxies: _this.proxies});
+            if (errors.length)
+            {
+                _this.emit('error', {message: 'Config errors',
+                    errors: errors});
+                throw new Error('Config errors: ' + JSON.stringify(errors));
+            }
         }
-        _this.start_web_socket();
         yield _this.prepare_database();
-        _this.db.main.stmt = {
-            proxy_remove: _this.db.main.prepare(`DELETE FROM proxy_hosts
-                WHERE host=$host`),
-            history: _this.db.main.prepare(`INSERT INTO request (port, url,
-                method, request_headers, request_body, response_headers,
-                status_code, status_message, timestamp, elapsed, response_time,
-                node_latency, country, timeline, super_proxy, proxy_peer,
-                username, content_size, context, success, out_bw, in_bw,
-                response_body, protocol, hostname)
-                VALUES ($port, $url, $method, $request_headers, $request_body,
-                $response_headers, $status_code, $status_message, $timestamp,
-                $elapsed, $response_time, $node_latency, $country, $timeline,
-                $super_proxy, $proxy_peer, $username, $content_size,
-                $context, $success, $out_bw, $in_bw, $response_body,
-                $protocol, $hostname)`),
-        };
+        if (cluster_mode.is_master())
+            _this.prepare_db_stmt();
         yield _this.logged_update();
         yield _this.create_socks_servers();
         yield _this.init_proxies();
-        if (_this.argv.www)
+        if (_this.argv.www&&cluster_mode.is_master())
         {
+            _this.start_web_socket();
             _this.www_server = yield _this.create_web_interface();
             _this.emit('www_ready', _this.www_server.url);
             if (!zutil.is_mocha())
@@ -2916,12 +3035,16 @@ E.prototype.start = etask._fn(function*mgr_start(_this){
                 console.log(boxed_line(bs));
             }
         }
-        _this.start_auto_update();
-        if (!is_win&&!is_darwin&&posix)
+        if (!is_win&&!is_darwin&&posix&&cluster_mode.is_worker())
             yield _this.start_ulimit();
         _this.is_running = true;
-        _this.perr('start_success');
-        _this.run_stats_reporting();
+        if (cluster_mode.is_master())
+        {
+            _this.start_auto_update();
+            _this.perr('start_success');
+            _this.run_stats_reporting();
+        }
+
     } catch(e){ ef(e);
         if (e.message!='canceled')
         {
