@@ -7,12 +7,10 @@ import etask from 'hutil/util/etask';
 import ajax from 'hutil/util/ajax';
 import {Modal} from './common.js';
 import $ from 'jquery';
-import util from './util.js';
+import {ga_event} from './util.js';
 import Pure_component from '../../www/util/pub/pure_component.js';
 import {If} from '/www/util/pub/react.js';
 import classnames from 'classnames';
-
-const ga_event = util.ga_event;
 
 class Notif_center extends Pure_component {
     constructor(props){
